+++ src/applications/mhv/secure-messaging/tests/fixtures/recipients.json
@@ -28,329 +28,5 @@
     "name": "***Mohammad_Fort Belovoir TG 200_APPT_SLC4%",
     "relationType": "PATIENT",
     "preferredTeam": true
-  },
-  {
-    "id": 6969633,
-    "name": "***RAJESH_APPOINTMENT CLINIC_SLC4^^^^^^",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 64158,
-    "name": "***RAJESH_PRIMARY_CLINIC_SLC4^^^^^^",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148501,
-    "name": "AARON_0704221254",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148510,
-    "name": "AARON_07151122_CLINIC_NEW",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148505,
-    "name": "AARON_0715220844_TEST_D",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148515,
-    "name": "AARON_0715221255_PCMM_LOGIN",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148523,
-    "name": "AARON_0719220915_LEGACY",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7225783,
-    "name": "ABDEL_12_07_PCMM",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7225784,
-    "name": "ABDEL_PCMM_455",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7208916,
-    "name": "Aaron-PCMM-102022-1400",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6832967,
-    "name": "BLUE ANCILLARY_TEAM",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 7123233,
-    "name": "CAMRY_PCMM RELATIONSHIP_05092022_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7173862,
-    "name": "CLINIC-Aaron-09-15-22-1354",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148248,
-    "name": "CLINIC_MOLLY_070722_MOHAMMAD_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6860253,
-    "name": "DavidPCMM_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 6864303,
-    "name": "David_PatientPreferencesTest1",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 7233648,
-    "name": "Demo_01_PCMM",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 46159,
-    "name": "GPK_TG_6",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 6955354,
-    "name": "GP_CLINIC NAME_TEST",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 6877504,
-    "name": "Geetha_SMClinician_Test1",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 7064072,
-    "name": "JIRA_23087",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 56219,
-    "name": "Liana's SLC4 PCMM",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6835136,
-    "name": "MOHAMMAD_CLINIC ASSOCIATION_BLUE ANCILLARY_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 21813,
-    "name": "MOHAMMAD_MULTIPLE TYPE ASSOCIATION_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6874047,
-    "name": "MOHAMMAD_OERR ASSOCIATION_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6866239,
-    "name": "MOHAMMAD_PCMM-R_PCP_TG_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 21015,
-    "name": "Mohammad Test_MRI_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 7223279,
-    "name": "PCMM ALEXANDRIA TG_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148491,
-    "name": "PCMM RELATIONSHIP_CAMRY_070722_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223296,
-    "name": "PCMM_ALGERIA_002_ABDEL_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7225787,
-    "name": "PCMM_ALGERIA_ABDEL_001_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223297,
-    "name": "PCMM_ALGERIA_ABDEL_003_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7225788,
-    "name": "PCMM_ALGERIA_ABDEL_004_SLC",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223298,
-    "name": "PCMM_ALGERIA_ABDEL_005_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223299,
-    "name": "PCMM_ALGERIA_ABDEL_006_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223300,
-    "name": "PCMM_ALGERIA_ABDEL_007_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223302,
-    "name": "PCMM_ALGERIA_ABDEL_009_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223281,
-    "name": "PCMM_ARMY VET TG_SLC4_MOHAMMAD",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223277,
-    "name": "PCMM_MOHAMMAD 12012022_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7223295,
-    "name": "PCMM_VIRGINIA_001_MOHAMMAD_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6866240,
-    "name": "RAJESH_PCMM ASSOCIATION_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6966391,
-    "name": "RAJESH_PCMM-R_DESIGNATED_WH_TG_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6864990,
-    "name": "RAJESH_TG(CARDIO-1)",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148520,
-    "name": "Rajesh-BLUE-CLINIC",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7217550,
-    "name": "Rajesh_Blue_Ancillary_Telaphone",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7148500,
-    "name": "SM-API-AARON-071422",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7183510,
-    "name": "SMPt_Reeng_API_intBTest",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6985315,
-    "name": "Steve's TG SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 7122538,
-    "name": "THOMAS CLINIC ASSOCIATION_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 7122544,
-    "name": "THOMAS_APPOINTMENT_MAY 09 2021_SLC4",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6827501,
-    "name": "VCCM_Test_01",
-    "relationType": "PATIENT",
-    "preferredTeam": false
-  },
-  {
-    "id": 6866272,
-    "name": "adfasfasfasdsd",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6864752,
-    "name": "b",
-    "relationType": "PATIENT",
-    "preferredTeam": true
-  },
-  {
-    "id": 6963734,
-    "name": "bhanu demo 1234",
-    "relationType": "PATIENT",
-    "preferredTeam": false
   }
 ]
